move bridge files to separate directory
